+++ src/main/java/frc/robot/poseestimator/observations/VisionObservation.java
@@ -1,5 +1,12 @@
 package frc.robot.poseestimator.observations;
 
 import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Pose3d;
 
-public record VisionObservation(Pose2d robotPose, double[] standardDeviations, double timestamp) {}
+public interface VisionObservation {
+
+	Pose3d getEstimatedPose();
+
+ 	double getTimeStamp();
+
+}
