+++ src/main/java/frc/robot/vision/rawdata/RawVisionData.java
@@ -1,8 +1,9 @@
 package frc.robot.vision.rawdata;
 
 import edu.wpi.first.math.geometry.Pose3d;
+import frc.robot.poseestimator.observations.VisionObservation;
 
-public class RawVisionData {
+public class RawVisionData implements VisionObservation {
 
 	protected final Pose3d estimatedPose;
 	protected final double timestamp;
